+++ src/guisim/GUIContainer.cpp
@@ -346,7 +346,7 @@ GUIContainer::setFunctionalColor(int activeScheme) const {
         }
         case 9: { // color randomly (by pointer)
             const double hue = (double)((long long int)this % 360); // [0-360]
-            const double sat = (((long long int)this / 360) % 67) / 100.0 + 0.33; // [0.33-1]
+            const double sat = (double)(((long long int)this / 360) % 67) / 100. + 0.33; // [0.33-1]
             GLHelper::setColor(RGBColor::fromHSV(hue, sat, 1.));
             return true;
         }
