+++ src/netedit/elements/demand/GNEVehicle.cpp
@@ -1807,7 +1807,7 @@ GNEVehicle::setColor(const GUIVisualizationSettings& s) const {
             case 29: { // color randomly (by pointer hash)
                 std::hash<const GNEVehicle*> ptr_hash;
                 const double hue = (double)(ptr_hash(this) % 360); // [0-360]
-                const double sat = ((ptr_hash(this) / 360) % 67) / 100.0 + 0.33; // [0.33-1]
+                const double sat = (double)((ptr_hash(this) / 360) % 67) / 100. + 0.33; // [0.33-1]
                 GLHelper::setColor(RGBColor::fromHSV(hue, sat, 1.));
                 break;
             }
